refactor(common,starknet_api): align tx hash logic to papyrus_common
