+++ crates/papyrus_common/src/transaction_hash_test.rs
@@ -4,7 +4,7 @@ use serde::{Deserialize, Serialize};
 use sha3::{Digest, Keccak256};
 use starknet_api::block::BlockNumber;
 use starknet_api::core::ChainId;
-use starknet_api::transaction::{Transaction, TransactionHash};
+use starknet_api::transaction::{Transaction, TransactionHash, TransactionOptions};
 use starknet_types_core::felt::Felt;
 
 use super::{
@@ -13,7 +13,6 @@ use super::{
     validate_transaction_hash,
     CONSTRUCTOR_ENTRY_POINT_SELECTOR,
 };
-use crate::TransactionOptions;
 
 #[test]
 fn test_ascii_as_felt() {
